\name{geco.enrichmentTest}
\alias{geco.enrichmentTest}
\title{geco.enrichmentTest}
\description{This function tests for gene set enrichements in a gene list using the hypergeometric test}
\usage{geco.enrichmentTest(gene.sets, mylist, possibleIds, sep=";", silent=F) }
\arguments{
\item{gene.sets}{List of gene sets to be tested}
\item{mylist}{List of selected genes}
\item{possibleIds}{List of genes that could have possibly been selected (e.g. genes present on the array)}
\item{sep}{required}
\item{silent}{required}
}
\value{required
}
\author{Eric Letouze}
\keyword{methods}
